backend/requirements.txt
fastapi
uvicorn
transformers
torch
librosa
psycopg2-binary
python-multipart